refactor: (#785) 코드 가독성을 위해 배너 관련된 코드 분리
